Jpa Database

+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -20,7 +20,7 @@ public class FileBasedDatabase implements Database {
   private final JsonService jsonService;
 
   @Override
-  public int save(Invoice invoice) {
+  public long save(Invoice invoice) {
     try {
       invoice.setId(idProvider.getNextIdAndIncrement());
       filesService.appendLineToFile(databasePath, jsonService.toJson(invoice));
@@ -32,7 +32,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(long id) {
     try {
       return filesService.readAllLines(databasePath)
           .stream()
@@ -57,7 +57,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> update(int id, Invoice data) {
+  public Optional<Invoice> update(long id, Invoice data) {
     try {
       List<String> allInvoices = filesService.readAllLines(databasePath);
 
@@ -91,7 +91,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(long id) {
     try {
       var allInvoices = filesService.readAllLines(databasePath);
 
@@ -112,7 +112,7 @@ public class FileBasedDatabase implements Database {
     }
   }
 
-  private boolean containsId(String line, int id) {
+  private boolean containsId(String line, long id) {
     return line.contains("\"id\":" + id + ",");
   }
 }
